requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
pycodestyle==2.10.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
